(bug 48145) Added time.Time.newFromIso8601
The time parser is currently not able to parse iso8601 in the form it is returned by
time.Time.iso8601. This function serves as a workaround to compensate for now.
We require this since we want the TimeValue data value type to store the iso8601
representation rather than all fields separately. The TimeValue's JSON will therefore
mainly consist of that iso8601 string from whom we have to create a new TimeValue in
the frontend.

Change-Id: I19372290cef7f6a768f687373cac6be59c77474b